Store seconds as Double type in VideoTime and increase progress bar refresh rate.

+++ iina/MPVChapter.swift
@@ -19,7 +19,7 @@ class MPVChapter: NSObject {
   var time: VideoTime
   var index: Int
 
-  init(title: String?, startTime: Int, index: Int) {
+  init(title: String?, startTime: Double, index: Int) {
     self.privTitle = title
     self.time = VideoTime(startTime)
     self.index = index
